requirements.txt
Flask==2.1.0
requests==2.22.0
click==8.1.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.1.2
zipp==3.8.0
python-dotenv